Documentation option and externalization of strings.
Major Update: Now commands support the "documentation option", i.e. a
special flag that can be typed as first argument of a command to show
its documentation. It is possible to create commands that are better
suitable for parsing arguments in a main() method, or to be inserted in
a CLI, through the Command.forMain() method and the Command.forCLI()
method.

Minor update: All message strings have been externalized into the
readycli.messages.properties resource bundle. At the moment two
languages are supported: english (the default language) and italian.