A command does not require a name anymore.
A command can be created by specifying a description and a usage string.
For a command that is used to parse arguments in a program, a name has
no meaning. On the other hand, a name is significant for a command
specified in a CLI. In this case the command name is used by the CLI to
address it when it is invoked.